utils: make this package build on non-linux platforms
Thie moves RunUnderSystemdScope and Syncfs to utils_linux and adds an
approximation of Syncfs on non-linux platforms.
